add filterFunc type
