+++ internal/chart/uninstall.go
@@ -4,10 +4,13 @@ import (
 	"fmt"
 
 	"k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/types"
 )
 
-func Uninstall(config *Config) error {
+type filterFunc func(unstructured.Unstructured) bool
+
+func Uninstall(config *Config, filterFunc ...filterFunc) error {
 	manifest, err := getManifest(config)
 	if err != nil {
 		return fmt.Errorf("could not render manifest from chart: %s", err.Error())
@@ -20,6 +23,10 @@ func Uninstall(config *Config) error {
 
 	for i := range objs {
 		u := objs[i]
+		if !fitToFilters(u, filterFunc...) {
+			continue
+		}
+
 		config.Log.Debugf("deleting %s %s", u.GetKind(), u.GetName())
 		err := config.Cluster.Client.Delete(config.Ctx, &u)
 		if errors.IsNotFound(err) {
@@ -36,4 +43,14 @@ func Uninstall(config *Config) error {
 		Namespace: config.Release.Namespace,
 	})
 	return nil
+}
+
+func fitToFilters(u unstructured.Unstructured, filterFunc ...filterFunc) bool {
+	for _, fn := range filterFunc {
+		if !fn(u) {
+			return false
+		}
+	}
+
+	return true
 }
